pyproject.toml
[project]
name = "engage"
version = "0.0.1"

[build-system]
requires = ["poetry==1.4.2"]

[tool.poetry]
name = "engage"
version = "0.0.1"
description = "City Council Engage-O-Tron"
authors = ["Dave Peck <davepeck@davepeck.org>"]

[tool.black]
target-version = ["py311"]
check = true
exclude = "/migrations/"

[tool.ruff]
exclude = ["migrations"]

[tool.isort]
known_first_party = "server"
profile = "black"

[tool.pyright]
include = ["server"]
typeCheckingMode = "basic"
useLibraryCodeForTypes = true

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
requests = "^2.28.2"
dj-database-url = "^1.3.0"
django-click = "^2.3.0"
pgvector = "^0.1.6"
psycopg = "^3.1.8"
openai = "^0.27.4"
pydantic = "^1.10.7"
pyhumps = "^3.8.0"
langchain = "^0.0.138"
django-stubs = "^1.16.0"
humanize = "^4.6.0"
beautifulsoup4 = "^4.12.2"
markdownify = "^0.11.6"
pdfplumber = "^0.9.0"
django-nonrelated-inlines = "^0.2"
lxml = "^4.9.2"
tiktoken = "^0.3.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.261"
